+++ src/libs/battle_interface/src/interface_manager/interface_group/sea_group.h
@@ -14,7 +14,7 @@ class BI_SeaGroup : public BI_BaseGroup
     {
     }
 
-    long Event(const char *pcEventName) override
+    int32_t Event(const char *pcEventName) override
     {
         return 0;
     }
